# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Management API",
        "description": "REST API for PT Telkomsel Inventory Management System. Authenticate via POST /api/v1/auth/login to receive a Bearer token.",
        "contact": {
            "email": "admin@telkom.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/borrowings": {
            "get": {
                "tags": [
                    "Borrowings"
                ],
                "summary": "List all borrowing transactions",
                "operationId": "97645ecbcd374d785fa7508755ce2b68",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "borrowed",
                                "returned",
                                "overdue"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated borrowing list"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Borrowings"
                ],
                "summary": "Create a new borrowing transaction (multi-item)",
                "operationId": "8b2c2751b9ae4f9b116bc267752f7a27",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "borrower_name",
                                    "borrow_date",
                                    "due_date",
                                    "items"
                                ],
                                "properties": {
                                    "borrower_name": {
                                        "type": "string",
                                        "example": "Budi Santoso"
                                    },
                                    "borrow_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-02"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-09"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Borrowing transaction created"
                    },
                    "403": {
                        "description": "Forbidden — Admin/Staff only"
                    },
                    "422": {
                        "description": "Validation error or insufficient stock"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/borrowings/{id}": {
            "get": {
                "tags": [
                    "Borrowings"
                ],
                "summary": "Get borrowing transaction detail",
                "operationId": "2f322ddf60e52d1d33f75a1ada834b8a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Borrowing detail with all items"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/borrowings/{id}/return": {
            "patch": {
                "tags": [
                    "Borrowings"
                ],
                "summary": "Process item return for a borrowing transaction",
                "operationId": "d13e997307d94908a2a199af248cae9b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "borrowing_detail_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "condition_after": {
                                                    "type": "string",
                                                    "enum": [
                                                        "good",
                                                        "lightly_damaged",
                                                        "heavily_damaged"
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return processed successfully"
                    },
                    "403": {
                        "description": "Forbidden — Admin/Staff only"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "List all categories",
                "operationId": "222626df48e0f5f8b98e79f320e987f4",
                "responses": {
                    "200": {
                        "description": "Category list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create a new category (Admin/Staff only)",
                "operationId": "973dcd9a9a46d5f57d4d646b3ebf9440",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Elektronik"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Perangkat elektronik kantor",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created"
                    },
                    "403": {
                        "description": "Forbidden — Admin/Staff only"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Data retrieved successfully"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter the token returned by POST /api/v1/auth/login",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}